agents/requirements.txt
# PariMarket agent dependencies
# Install with:  uv pip install -r requirements.txt
# Or from project root:  uv pip install -e ".[dev]"

google-adk>=1.0.0
google-generativeai>=0.8.0
web3>=7.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
mcp>=1.2.1